Refactored auth tests
